+++ allennlp/training/callbacks/wandb.py
@@ -1,3 +1,4 @@
+import logging
 import os
 from typing import Optional, Dict, Any, List, Union, Tuple, TYPE_CHECKING
 
@@ -13,13 +14,18 @@ if TYPE_CHECKING:
     from allennlp.training.trainer import GradientDescentTrainer
 
 
+logger = logging.getLogger(__name__)
+
+
 @TrainerCallback.register("wandb")
 class WandBCallback(LogWriterCallback):
     """
     Logs training runs to Weights & Biases.
 
     !!! Note
-        This requires the environment variable 'WANDB_API_KEY' to be set.
+        This requires the environment variable 'WANDB_API_KEY' to be set in order
+        to authenticate with Weights & Biases. If not set, you may be prompted to
+        log in or upload the experiment to an anonymous account.
 
     In addition to the parameters that `LogWriterCallback` takes, there are several other
     parameters specific to `WandBWriter` listed below.
@@ -50,7 +56,9 @@ class WandBCallback(LogWriterCallback):
         files_to_save: Tuple[str, ...] = ("config.json", "out.log"),
     ) -> None:
         if "WANDB_API_KEY" not in os.environ:
-            raise ValueError("Missing environment variable 'WANDB_API_KEY'")
+            logger.warning(
+                "Missing environment variable 'WANDB_API_KEY' required to authenticate to Weights & Biases."
+            )
 
         super().__init__(
             serialization_dir,
@@ -72,6 +80,7 @@ class WandBCallback(LogWriterCallback):
             project=project,
             config=Params.from_file(os.path.join(serialization_dir, "config.json")).as_dict(),
             tags=tags,
+            anonymous="allow",
         )
 
         for fpath in self._files_to_save: